requirements/base.txt
# Django
Django>=5.0,<6.0

# Database
psycopg2-binary>=2.9.0

# Configuration
python-decouple>=3.8
dj-database-url>=2.1.0

# Forms
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7

# Image handling
Pillow>=10.0.0

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0

